requirements.txt
Django==5.2.10
django-tenants==3.9.0
psycopg[binary]>=3.2.1,<3.3
python-dotenv==1.0.1
